final/pyproject.toml
[tool.poetry]
name = "final"
version = "0.1.0"
description = ""
authors = ["Alexandre Jeanne <alex.eco372@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
reach = "^3.4.7"
rich = "^12.4.1"
pyserde = {extras = ["all"], version = "^0.7.3"}
typer = "^0.4.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
